+++ pages/course/[courseId]/games/index.tsx
@@ -4,7 +4,6 @@ import type { NextPage } from "types";
 
 const Games: NextPage = () => <GamesPage />;
 Games.authRequired = true;
-Games.proRequired = true;
 
 export default Games;