+++ src/pages/GamesPage/GamesPage.tsx
@@ -3,6 +3,8 @@ import Select from "@/atoms/Select";
 import TextInput from "@/atoms/TextInput";
 import SEO from "@/helpers/SEO";
 import { getElementsVals } from "@/helpers/getElementsVals";
+import useProStore from "@/stores/proStore";
+import Link from "next/link";
 import { useRouter } from "next/router";
 import { useState } from "react";
 
@@ -12,12 +14,20 @@ type Game = "MATCHING-GRID" | "ALU-BOT" | "QUIZ";
  * Interface to select which game to play (and the settings for that game)
  */
 export default function GamesPage() {
+  const isPro = useProStore((state) => state.isPro);
   const [game, setGame] = useState<Game>("MATCHING-GRID");
+  const [showProRequired, setShowProRequired] = useState(false);
   const router = useRouter();
   const { courseId } = router.query;
 
   const play = (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
+
+    if (!isPro) {
+      setShowProRequired(true);
+      return;
+    }
+
     const pathname = `/course/${courseId}/games/${game.toLowerCase()}`;
     const { flashcardsType } = getElementsVals(e.target as HTMLFormElement, [
       "flashcardsType",
@@ -65,7 +75,18 @@ export default function GamesPage() {
         description="Play fun games with flashcards at Alu Learn. Matching games, chatbot games, quiz games, and more are available"
       />
       <div className="mt-28 max-w-2xl mx-auto">
-        <h1 className="text-4xl font-bold text-center mb-2">Games</h1>
+        <h1 className="text-4xl font-bold text-center">
+          Games {!isPro && "(Pro Only)"}
+        </h1>
+        {!isPro && (
+          <p className="text-center mt-1 mb-3">
+            Only Alu Pro users can play games with their flashcards.{" "}
+            <Link href="/pro">
+              <a className="text-blue-500">Start your 7-day free trial today</a>
+            </Link>
+            .
+          </p>
+        )}
         <form onSubmit={play}>
           <Select
             label="Game"
@@ -113,6 +134,17 @@ export default function GamesPage() {
           <Button className="mt-4" type="submit" block>
             Play!
           </Button>
+          {showProRequired && (
+            <p className="text-center mt-3">
+              Only Alu Pro users can play games with their flashcards.{" "}
+              <Link href="/pro">
+                <a className="text-blue-500">
+                  Start your 7-day free trial today
+                </a>
+              </Link>
+              .
+            </p>
+          )}
         </form>
       </div>
     </>